+++ benchmark/data_structure/string.cc
@@ -35,7 +35,7 @@ class StringTestSuite {
   }
 
   /** Construct + destruct in a loop */
-  void ConstructDestructTest(int count, int length) {
+  void ConstructDestructTest(size_t count, int length) {
     char *data = (char *) malloc(length + 1);
     data[length] = 0;
     memset(data, 1, length);
@@ -51,7 +51,7 @@ class StringTestSuite {
   }
 
   /** Construct in a loop, and then destruct in a loop */
-  void ConstructThenDestructTest(int count, int length) {
+  void ConstructThenDestructTest(size_t count, int length) {
     char *data = (char *) malloc(length + 1);
     data[length] = 0;
     memset(data, 1, length);
